Make blockies reusable
