+++ blockies/src/main/java/pm/gnosis/blockies/BlockiesImageView.kt
@@ -8,17 +8,13 @@ import android.util.AttributeSet
 import android.widget.ImageView
 import pm.gnosis.model.Solidity
 
-open class BlockiesImageView(context: Context, attributeSet: AttributeSet) : ImageView(context, attributeSet) {
-    private val canvasPaint = Paint().apply { style = Paint.Style.FILL }
-    private var dimen = 0.0f
-    private var offsetX = 0.0f
-    private var offsetY = 0.0f
-    private val path = Path()
+open class BlockiesImageView(context: Context, attributeSet: AttributeSet?) : ImageView(context, attributeSet) {
 
     private var blockies: Blockies? = null
+    private var painter: BlockiesPainter = BlockiesPainter()
 
-    fun setAddress(address: Solidity.Address) {
-        blockies = Blockies.fromAddress(address)
+    fun setAddress(address: Solidity.Address?) {
+        blockies = address?.let { Blockies.fromAddress(it) }
         invalidate()
     }
 
@@ -29,44 +25,10 @@ open class BlockiesImageView(context: Context, attributeSet: AttributeSet) : Ima
 
     override fun onMeasure(widthMeasureSpec: Int, heightMeasureSpec: Int) {
         super.onMeasure(widthMeasureSpec, heightMeasureSpec)
-        dimen = Math.min(measuredWidth, measuredHeight).toFloat()
-        offsetX = measuredWidth - dimen
-        offsetY = measuredHeight - dimen
-        path.reset()
-        path.addCircle(offsetX + (dimen / 2), offsetY + (dimen / 2), dimen / 2, Path.Direction.CCW)
-        path.close()
+        painter.setDimensions(measuredWidth.toFloat(), measuredHeight.toFloat())
     }
 
     private fun drawBlockies(canvas: Canvas, blockies: Blockies) {
-        canvas.save()
-        canvas.clipPath(path)
-        canvasPaint.color = blockies.backgroundColor
-        canvas.drawRect(
-            offsetX, offsetY, offsetX + dimen, offsetY + dimen,
-            canvasPaint
-        )
-
-        val scale = dimen / Blockies.SIZE
-        val main = blockies.primaryColor
-        val sColor = blockies.spotColor
-
-        for (i in blockies.data.indices) {
-            val col = i % Blockies.SIZE
-            val row = i / Blockies.SIZE
-
-            canvasPaint.color = if (blockies.data[i] == 1.0) main else sColor
-
-            if (blockies.data[i] > 0.0) {
-                canvas.drawRect(
-                    offsetX + (col * scale),
-                    offsetY + (row * scale),
-                    offsetX + (col * scale + scale),
-                    offsetY + (row * scale + scale),
-                    canvasPaint
-                )
-            }
-        }
-
-        canvas.restore()
+        painter.draw(canvas, blockies)
     }
 }
